+++ apis/r/NEWS.md
@@ -13,3 +13,4 @@
 * All R6 classes' `create()` method now return `self` rather than nothing
 * Fixed calculating of relative paths when 1 of the URIs contains the `file://` prefix
 * Added `PlatformConfig` and `SOMATileDBContext` classes to handle SOMA and TileDB configuration
+* Add Seurat outgestors for `SOMAExperimentAxisQuery` objects